move cpe/error-utils to utils/error
