+++ packages/preview-middleware-client/src/cpe/changes/service.ts
@@ -20,7 +20,7 @@ import type Event from 'sap/ui/base/Event';
 import type FlexCommand from 'sap/ui/rta/command/FlexCommand';
 import Log from 'sap/base/Log';
 import { modeAndStackChangeHandler } from '../rta-service';
-import { getError } from '../error-utils';
+import { getError } from '../../utils/error';
 
 interface ChangeContent {
     property: string;
